Fix: style touchups
